Fix auth-related issues
Changed regex in users-router.js to make password validation more effective.
Also needed to fix an error in my seed data where a user id was assigned to
a category but an associated activity didn't have that same user id. I also needed to change the way a DELETE request was handled so that a deleted category would also have all of its associated activities removed.